progressbar: fix regression with hidden bars
The documentation of the progress bar states that the progress bar
will output only the label of the bar if it is outputting to a file
that is not a tty. However, this was broken a few versions ago,
without the documentation being adjusted, and even a test added.

Restore this behaviour so we follow the documentation again, and
adjust the test to match. Fixes #1138.