Added examples of default to boolean flags section (#1697)
I've found myself a few times double checking what the default option is when a boolean flag is use i.e. does `default=False` refer to the first flag or the second flag.  Thought adding an example of the calls without any flags to the docs would make this a bit clearer.