Allow a callable default to have its own string representation
As it was before 78a62b37c956e71b32689a0d5c75a684aa9ab56d.

A callable can also be an object with __call__ and __str__ defined, in
which case it makes sense to use it. Don't try too hard to show something
else instead, because the developer most likely knows what he (or she) is
doing.

https://github.com/pallets/click/issues/2099