Store raw help string for commands
Some tools, such as sphinx-click [1], may wish to access the full help
string for a command including any text after the form feed character
(which indicates truncation [2]). Make this possible by storing the
help string without truncation and instead truncate when we use this
string (i.e. truncate on load, not on store).

[1] https://github.com/click-contrib/sphinx-click/issues/56
[2] https://click.palletsprojects.com/en/latest/documentation/#truncating-help-texts

Signed-off-by: Stephen Finucane <stephen@that.guru>